Ensure that we can serialize app charge resources
This is necessary because the charge resources have an attribute called
'test'. In ActiveModel's serializer it does a __send__ call to each
attribute to get it's value. Since Kernel provides a private #test method
that one will be favoured over our 'test' attribute and serialization will
raise an error.

This deletes Kernel#test when the class is defined so there's never a
conflict.